[Feat] Kafka 기능 수정
1. 프로듀서/컨슈머 분리
2. 불필요 서비스 삭제(user-service)

+++ order-service/src/main/java/com/example/orderservice/dto/OrderItemDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -14,6 +16,6 @@ public class OrderItemDto {
     private Long productId;
     private String productName;
     private Integer quantity;
-    private Double price;
-    private Double subtotal;
+    private BigDecimal price;
+    private BigDecimal subtotal;
 }
